print unique prefixes and their entry value
Prints the prefixes in the trie that are unique, as in, they only have
one entry value at the end of the chain. Print the entry value as well.

This is the least amount of consecutive prefix of a string that uniquely
determines the entry value.

Optimize: skip parts of a string prefix when there is no other
possibility of continuation. As in: skip nodes with only 1 child untill
there are more children again.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>