consider buffer overflow case and handle trailing chars
The trailing characters in one buffer read are now copied to the front
of the buffer and the new characters are appended to the characters from
last iteration. Possibly push up the vector operations that are not
required anymore since they are already processed.

Buffer overflow happens when lines are longer than the buffer, which can
never be processed correctly. Throw an error in this case since it's
unrecoverable.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>